Forecast for Vehicles Sold
The forecast is in its final stage, where I have to add the seasonality back to the arima model. 
I have identified that the time series is best describe with an ARIMA(2,1,3) based on AIC. 

However, I cannot add the final part. It works if I remove the first difference and just run a ARMA(2,3).